pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-home-app"
version = "0.1.0"
description = "Privacy-first local grocery receipt intelligence system"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pydantic>=2.7,<3",
  "pydantic-settings>=2.2,<3",
  "Pillow>=10,<12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9",
  "ruff>=0.15,<0.16",
  "black>=25.11,<26",
  "mypy>=1.19,<2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
